Feature watch

+++ front/github_clone_front/src/services/RepositoryService.js
@@ -4,12 +4,12 @@ const create = (repoData) => {
     return api.post("repository/", repoData);
 }
 
-const starr_it = (username, repoName) => {
-    return api.post(`repository/starr_it/${username}/${repoName}/`);
+const starr_it = (username, repoName, owner_username) => {
+    return api.post(`repository/starr_it/${username}/${repoName}/${owner_username}/`);
 }
 
-const unstarr_it = (username, repoName) => {
-    return api.delete(`repository/unstarr_it/${username}/${repoName}/`);
+const unstarr_it = (username, repoName, owner_username) => {
+    return api.delete(`repository/unstarr_it/${username}/${repoName}/${owner_username}/`);
 }
 
 const get = (username, repoName,logged_user) => {
@@ -100,5 +100,11 @@ const fork = (ownerUsername, repoName, data) => {
     return api.post(`repository/fork/${ownerUsername}/${repoName}/`, data);
 }
 
+const saveWatchPreferences = (ownerUsername, repoName, data) => {
+    return api.post(`repository/watch/${ownerUsername}/${repoName}/`, data);
+}
+
 export default { unstarr_it,starr_it,getAllStaredUserRepos,getIsUsersRepo,getAllQueryRepos,create, get, getOwner, getRootContent, getFolderContent, update, deleteReposiory, getAllUserRepos, getFile,
-    editFile, deleteFile, createFile, inviteCollaborator, respondToInvitation, getInvitation, getCollaborators, removeCollaborator, changeRole, transfer, fork };
+    editFile, deleteFile, createFile, inviteCollaborator, respondToInvitation, getInvitation, getCollaborators, removeCollaborator, changeRole, transfer, fork,
+    saveWatchPreferences
+};
